Finished porting fitter as task
