+++ AtReconstruction/AtFitter/AtFitter.h
@@ -12,6 +12,7 @@
 
 class AtDigiPar;
 class AtTrack;
+class AtFittedTrack;
 class FairLogger;
 class TBuffer;
 class TClass;
@@ -28,8 +29,7 @@ class AtFitter : public TObject {
 public:
    AtFitter();
    virtual ~AtFitter();
-   // virtual std::vector<AtTrack> GetFittedTrack() = 0;
-   virtual genfit::Track *FitTracks(AtTrack *track) = 0;
+   virtual std::unique_ptr<AtFittedTrack> ProcessTracks(std::vector<AtTrack> &tracks) = 0;
    virtual void Init() = 0;
 
    void MergeTracks(std::vector<AtTrack> *trackCandSource, std::vector<AtTrack> *trackJunkSource,
